Fixed USB_GetHIDReportSize() returning the number of bits in the specified report instead of bytes.
Moved the USB_GetHIDReportItemInfo() calls into the main report item passing loop in the *HostWithParser demos - it is fast enough not to effect performance, and avoids duplicate code.

Make Webserver project report the LUFA version as part of the HTTP header.

+++ LUFA/Drivers/USB/Class/Host/HIDParser.h
@@ -215,7 +215,7 @@
 			typedef struct
 			{
 				uint8_t                      ReportID; /** Report ID of the report within the HID interface */
-				uint8_t                      ReportSizeBits[3]; /** Total number of bits in each report type for the given Report ID,
+				uint16_t                     ReportSizeBits[3]; /** Total number of bits in each report type for the given Report ID,
 				                                                 *  indexed by the \ref HID_ReportItemTypes_t enum
 																 */
 			} HID_ReportSizeInfo_t;
